Fix erd generate errors

+++ src/backend/database/migrations/2024_03_21_121825_add_message_and_trace_columns_to_company_jobs_table.php
@@ -23,9 +23,11 @@ return new class extends Migration {
      * @return void
      */
     public function down() {
-        Schema::table('company_jobs', function (Blueprint $table) {
-            $table->dropColumn('message');
-            $table->dropColumn('trace');
-        });
+        if (Schema::hasTable('company_jobs')) {
+            Schema::table('company_jobs', function (Blueprint $table) {
+                $table->dropColumn('message');
+                $table->dropColumn('trace');
+            });
+        }
     }
 };
